SSH brute force protection
To block a SSH brute force attack, we just need to slow down the
flow of requests. We can do this by rate-limiting requests to SSH
with iptables. The benefit of using iptables to block SSH attacks
is you don’t need any added software so we can easily support this
solution.

This change will block an IP if it attempts more than 3 connections
per minute to SSH.

DocImpact
Change-Id: I0f452c8b0a808789aa4c2cd85d1d00556b210a39
Closes-Bug: #1540073