Don't use 'include' syntax in tasks
Using Include and Require functions for class
declaration allows the class to me declared many
times within a single manifest without any
duplicate decalration error. But the cost of this
is that the file and line where the class is declared
is not saved in the catalog object.

The Coverage parser cannot determine if this class was
declared in the particular task and skips these classes
making the coverage report in the Noop tests much less
relevant.

Change-Id: Ie5d5ced4a9225bda5b0bf6c2a733c0d44c48118c