Remove extra loops from upload_cirros script
This change removes the extra loops from the upload_cirros script that
were added while attempting to debug Bug 1556068.  These extra loops
could cause the task to exceed the task timeout and ultimately did not
address the main issue. This change cleans up the extra code that was
added for debugging purposes now that the previous bug has been closed.

Change-Id: I26cfe1d0f6f2f6d8fb2e9fbba2cd964574c7bbef
Closes-Bug: #1563981