Set default iptables INPUT policy to DROP
It was observed that on Fuel master node the default firewall INPUT
policy is set to 'ACCEPT' rather than to 'DROP'.
This leads to exposure of unnecessary services over a potentially
untrusted networks.

This patch updates default firewall INPUT policy to 'DROP'. Also it adds
user-defined chains which are not purged (and thus preserved between
iptables.pp applications).

Change-Id: Ia9ab6d019be81aebcf5eaba25336e6f19b2c6a1a
Partial-Bug: #1524750
Depends-On: I57e9f58c6bad32b23b179499f0514edf5357bd31
Signed-off-by: Maksim Malchuk <mmalchuk@mirantis.com>