Purge deleted /etc/hosts entries for deleted ndoes
Deleted nodes now will get stored in a file
/etc/hiera/deleted_nodes.yaml and be inserted into
the hiera hierarchy.

host entries for any records present in deleted_nodes.yaml
will be purged when evaluating hosts task.

The source YAML for deleted nodes is generated on master
because nodes that are deployed from scratch or from Error
state will not be able to identify deleted nodes due to
certain orchestration limitations.

Change-Id: I01db215e77a3532a6fa7bf46ab7e20e281e8c165
Closes-Bug: #1513401