Generate vips: puppet4 and latest 3.8 support
The generate_vips function is refactored to work in pair
with the create_resources functions. It's much better for testing
and debugging.
Input data of the vurual ip definition is split into many parameters
and the generate vips function is tought to support the changed
interfacs as well as to work with Puppet4's undef values.

Related-Bug: 1586480
Change-Id: I317dc6f889074fe89f1780b8edb840c32fa860c9