Add new flag migrate process finishing.
For qa test it is necessary to determine when migration process finished.
For this purposes /not_ready was used. But it has some disadvantages, it
covers not whole migration process. And there was possibility to mistake.
So was added new  flag which appears after finishing of migration. Old
was left because of old tests. Both of them are used only for testing
propouses and don't affect migration process.

Partial-Bug: #1576276
Change-Id: I49b62dc743111cb4c41489c528e48f8d56c2f9df