Limit number of Keystone workers to 5 on master node
Currently number of Keystone workers is set based on number of CPU
threads. Each worker starts 5 connections to PostgreSQL, while we limit
number of possible connections to 100. This leads to situation where
Keystone fails to connect to the database, resulting in authentication
failure when Fuel command line client is used.

Change-Id: I94d5e995e78cdcb25d2ea3593701456526109ead
Closes-Bug: 1572498
Signed-off-by: Maksim Malchuk <mmalchuk@mirantis.com>