Make keystone session stickiness optional
As part of Bug 1527717, we added in support of session stickiness for
keystone to allow for federation support in keystone. This introduced
performance issues with keystone and should not be configured by
default. This change disables the stickiness by default but allows for
it to be enabled by adding a 'federation' key to the keystone hiera
hash.

DocImpact: Keystone HAProxy session stickiness can be enabled by adding
a 'federation' key to the keystone hiera configuration. By default
session stickiness is disabled.

Change-Id: Id43446a9f923417b8fa140ffcfddb6797f1b4e48
Related-Bug: #1527717
Closes-Bug: #1582202
(cherry picked from commit 8ebe5f15474319fafe2cdae2fc74ee98d63dcb80)