Update keystone noop tests for wsgi files
Change I03a7a057cae0bf86331191faf47ec394487150a7 is updating the source
path of the wsgi files used to configure keystone under apache. This
change updates the noop tests to just check that the files exist and
configured in the catalog but to exclude the path information.  Once the
change in upstream gets merged, we could add the path information back
in but fuel really shouldn't care where it lives so long as they are
configured.  The upstream tests should be relied on to ensure the
source/destination paths are managed correctly.

Change-Id: I9969185306c77254e781b2c52a6b9bededf40a98
Related-Bug: #1539201