Increase updatedb timeout
The current value of 60 seconds can be exceeded if the host that's being
deployed on is overloaded. We have seen (in CI) that the combined time
between fact loading and the exec can exceed 60 seconds if the system is
overloaded on IO or CPU.  This change increases the timeout slightly to
account for possible hardward load issues and the execution of this task.

Change-Id: I713bedb1a3b1d1aafb5c66f92542e4a57d659e44
Closes-Bug: #1580615