Run pki setup when ceph is enabled
Previously we had removed the PKI setup as part of the keystone since it
is deprecated upstream. Unfortunately we still rely on it for ceph
deployments with radosgw. This change puts it back in to resolve the
blocking failure with ceph. The PKI setup is still deprecated and will
need to be addressed with an alternative configuration for radosgw.

Change-Id: I9e9fd454e4be358a622e98fd7335ec937e0a3476
Closes-Bug: #1584851
Related-Bug: #1583557