Clean RabbitMQ definitions file
Added a script to clean up the rabbitmq definitions file which is
being generated during the post-deployment phase of a deployment.
Previosly the definitions file contained all RPC queues presented in
RabbitMQ after the successful deployment. These queues are one-off, and
their restoration via OCF-script leads to growth of unconsumed messages,
which, in turn, lead to RabbitMQ slowness and malfunctions.

This is a rework of the previously reverted change
Ie3f3194b3eaaa8dc28d1cd3442090b586e177306 which caused deployment
failures in some scenarios involving partial update. This happened
because the script was delivered in a separate package which has not
been updated during the partial update. Now the script is
delivered in the puppet's directory instead. Also introduced
additional regexp-matching logic to instantly remove all the RPC queues,
which were left in the previous version of the fix.

Change-Id: Ib2f8237692d7dab47b83ac886ff370686a5d22fe
Closes-Bug: #1497961