Do not install glare murano config under UCA
UCA does not provide the murano glance artifacts plugin so we need to
exclude it's configuration when deploying UCA.

DocImpact: UCA with murano does not install and configure glare as it is
not available with UCA packages. It will continue to use murano for the
artifact repository when UCA is used.

Change-Id: I356600cf410aaf0ce32f25ff18642e135b7380e9
Closes-Bug: #1586141
(cherry picked from commit 874b4350b4e3cb4a19aedb1d09f4fbcc1fec6f2d)