nova: configure file injection by the means of config drive
We've seen a number of problems with qemu-nbd (from files not being
injected randomly to kernel panics), and while libguestfs is a
better solution here it does not work out of box on Ubuntu. Neither
of these can be used for file injection when instance ephemeral
drives are stored in Ceph.

Disabling of file injection in Nova libvirt drive allows us to unify
the settings for Ubuntu vs CentOS and file vs Ceph-based ephemerals.
In this case file injection can still be performed by the means of
config drive + cloud-init within images.

Upstream Nova / Devstack explicitly disabled file injection in favor
of using config drives in:

9ce99a44cf85e431227536e2251ef05b52e61524

and

I2388ef0df12a6289b619bfaf30cb952fcc48ef41


DocImpact

File injection is no longer performed on the compute node because it's
slow and error-prone. Instead files to be injected are placed on a
config drive, which is automatically created for every instance. It's up
to cloud-init or a similar mechanism within the image itself to perform
injection on instance boot.

Closes-Bug: #1467860
Closes-Bug: #1556819
Related-Bug: #1467579
Related-Bug: #1493767

Change-Id: Ie46aa3f48d62b7500a7e326348b35573b3262641