Disable rabbitmq cluster partition handling
Currently autoheal functionality is a subset of OCF script
functionality (autoheal is concerned only with network partitions, OCF
also handles all sort of cases where rabbitmq becomes stuck). But as
both of this functions cause stopping and starting of rabbits, it
results in unwanted interference - as we've observed in 1585128, where
it triggered bug in autoheal code.

Even if rabbitmq bug will be fixed, it makes no sense to use both
autoheal and OCF script (actually, when clustered rabbit will stop
getting stuck during network split, autoheal will be as good as OCF
script).

Change-Id: I38fbf6d50e9aed35f3c5e3bc1e17de7001304706
Closes-Bug: 1585128