Unified (one) error message for all invalid requests
It is possible based on the server error messages, to enumerate
directories and files on the application. This commit change the
default Nginx error messages to Unified one message with 200 (OK)
HTTP status code. For 404 (NOTFOUND) error we should preserve the
status code but provide the same Unified message because the
Package managers (in the Debian/Ubuntu at least) should know the
status of the file existence in the repos.

Change-Id: I6b996c5b1f8a1972379365571cf1f261e61b9ac5
Closes-Bug: #1585176