Always wait for database backend
Previously when using the detached-database functionality, the various
tasks that require the database task on the primary controller will run
immediately after the primary controller database task has executed.
Because the wait for the mysql backend was included within the enabled
if block, when detached db is used the database task finishes
immediately without the database actually being ready. This change
takes the mysql backend wait code and puts it in it's own class that
can be included anywhere the mysql backend wait needs to occur.
Additionally this code can be leveraged as a task if necessary.

Change-Id: Ia1d45f45cbbd91b7c6cacd4aa46dae0a92e4156e
Closes-Bug: #1588387
(cherry picked from commit 39ba427e5fd03f8b69fe2109c0bbb4ee4748a70f)