Change configuration symlink expression to look at cluster status
There is a cases for node, when $.nodes.uid doesn't changed but
task should be ran. One of this cases is cluster reinstallation
after successful deploy, cause in this case all $.nodes metadata
stays the same as it were after cluster deployment, but nodes are
really destroyed and reprovisioned and as a result tasks must be
ran again. But cluster status in this case _will_ be changed, so
it is better to use it in expression.

Change-Id: I98b48db03ab87ce233ebf74dacb56488c3a16435
Closes-Bug: #1588320