Error if librarian-puppet fails
This change checks the return code status of librarian puppet and
stops the script execution if there is anerror.  Without this change
if librarian-puppet encounters an error, the script would continue
and not all the modules may be available. This is problematic for
the build systems that are using this script to pull down the modules.

Change-Id: Ic7277deb957072abcae9a8639196938b24288cfa
Closes-Bug: #1588895