Puppet4 support: refactor vips
The generate_vips function is refactored to work in pair
with the create_resources functions. It's much better for testing
and debugging.
Input data of the vurual ip definition is split into many parameters
and the generate vips function is tought to support the changed
interfacs as well as to work with Puppet4's undef values.

Change-Id: If865f39a3d1c22efa6df8035a4e92004fcf2ffb6
Related-Bug: 1586480