Puppet4 support: osnailyfacter
* Multiple fixes in the spec functions
* Disabled tests for wait_for_nova_backends
  They are unstable and very dependant on
  Puppet version. They should be rewritten later.
* Removed setvar and store function.
  They are not used anywhere and may
  be harmful if used.

fetch_value function

The functions for extracting values from a nested structure
have a long and funny history. Currently this function is
named "dig" and is used in the Fule linrary by this name.
But in Puppet 4.5 another incompatible "dig" function was
introduced and the original "dig" in the stdlib was renamed
to dig44.

The dig function from the stdlib doesn't support :undef values
and has problems with other values too. This function is
based on the "dig44" but has some improvements needed for
Puppet4.5+.

Change-Id: I9f7a07a582a0d67b5db570378b856cf0f4c7c48e
Related-Bug: 1586480