Disable all the hiera yamls for globals except for astute.yaml
We need to ensure that globals uses only astute.yaml to generate
its helper variables. Othewise, we will have an issue on the second
deployment if we modify astute.yaml and re-run the deployment

For globals.pp disable_globals variable is used to redirect hiera
into non-existing files and thus skip their inclusion.

Change-Id: I91ab492b44c137c476722f65de3dd99b2dce988a
Closes-bug: #1590375