Improve reliability of upload_cirros.rb
Previously we did not make sure the glance image was in an active state
before exiting the upload_cirros.rb script. This change checks that the
image upload was successful and the image is reporting active before
returning success. Additionally this change increases the number of
times we attempt to upload the image within the script and increases the
task retries as well. We will attempt to cleanup the glance image
between uploads as well if there are any failures.

Change-Id: I17187cc6ec596a2631b958a9636ca984128b8cd7
Partial-Bug: #1589433