Do not add upstream to resolv.conf on master
In order to not break local name resolution for the fuel environment, we
should rely on dnsmasq to query the upstream servers for us. Because of
this, the resolv.conf on the master should only point to the cobbler
dnsmasq instance which is configured with the upstream hosts to use if
the local resolution fails.

Change-Id: I757d89173d98dda419d21ef6132d59f186a97782
Closes-Bug: #1567957