Fix database tasks when detached database is used
This patch essentially reverts change sets
Ia1d45f45cbbd91b7c6cacd4aa46dae0a92e4156e and
I38ad9d5fd63dda374274858632710c121d721451 which were created because
of the regression introduced by
Icc27b6701726afb98df71050c344c6cd9525301e. But we are also reducing
the cross-depends to only match the primary-database and database
tasks to prevent possible loops.

We can only wait on the mysql backend if it's configured and we don't
configure the mysql haproxy backend if mysql is not enabled on a node.
Additionally we need to wait until the database is actually setup before
running any of the db tasks.

Change-Id: I4b9bc29b4134872908af1900f6a3e5ccbdb9e392
Closes-Bug: #1591256