Stop looking for master once latest seqno found
This change updates how we look for a master by stopping once we have
found a service with the largest seqno. Previously if all servers
had the same seqno then we would return the last server as a master
rather than the first. This had the side effect that during bootstrap
when the first server was the original master, it would be demoted for
each new server that was added to the group.  This should prevent the
ocf script from shuffling masters if they have the same seqno. We will
always pick the first server rather than the last.

Change-Id: Iacbd2e2ec403985a1ff52880669b1bec62dbbaba
Closes-Bug: #1592401
Related-Bug: #1592819
(cherry picked from commit 26f1690d10a5c8f401033dff00166c14bf77d4ac)