Fix incorrect mysql split brain determination
As part of the bootstrap process for galera, we can get in to a
state where the original master node thinks it gets in a splitbrain
condition after the the other two nodes bootstrap each other. This
change adds a check before the split brain determination to make sure
that the node isn't already current based on the GTID/SEQNUM.

What happened is node 1 is created and has the --wsrep-new-cluster
parameter, node 2 syncs from node 1. Node 3 joins and syncs from node 2.
Then when they try and fully form a quorum, we pull the GTID from all
the nodes and since they are all identical, the last node is considered
a canidate to be a master. Without this change, we would then check if
the node is a new cluster (has the --wsrep-new-cluster option) and then
return a generic error which would mean the node would try and resync.
With this change, we check to make sure that we aren't already
consistent before going down the path of assuming it's a new cluster.

Change-Id: Iacbd2e2ec403985a1ff52880669b1bec62dbbaba
Closes-Bug: #1592401