Remove erlang/rabbit thread pool calculation
Proper version of this calculation is implemented upstream -
https://github.com/rabbitmq/rabbitmq-server/issues/151

And removed version of the code is actually harmful, as it's using
physicalprocessorcount (i.e. number of CPU sockets) for
calculations. So on a 2 CPU/48 thread system there it was 30 threads
instead of 768 calculated upstream.

I've decided that it's easier to completely remove this code instead
of duplicating current formula in upstream and providing a way to
override this value through hiera - just not worth a hassle.

Change-Id: I415d446af0a822d2a5ce3478fd9db1dd0f13e115