Remove outdated stuff from Murano modular manifest
This patch removes unneeded backend checker for Murano - it was used
for uploading murano package, but now it is under different task on
post_deployment stage, hence we don't need it anymore.

Also backend checker is unneeded for CFAPI - there are no any operations
with keystone or murano for this modular task.

Also patch removes deprecated parameters because of upstream puppet
murano patch [1]

[1] I5ff106ba136994631c9eb69111a0b641891c9bb1

Change-Id: I4c81ccc889e4e4981866585e043c5123cc77f750