Use memcache for keystone_authtoken
This change updates where we calculate the memcache server information
to the globals hiera data and updates all the openstack services to
use the heira memcached_servers key for consistency. Additionally all
api services have been updated to configure the memcached servers for
the keystone_authtoken middleware configuration to improve the
responsiveness of the service.

Change-Id: Id1034e22d79c3ea6b25575d9bcf8e8750a02365d
Closes-Bug: #1597512