Check if upstream rabbitmq ocf script is installed
This change adds a check to the fuel rabbitmq ocf script to ensure we
are returning the correct return code. Because the rabbitmq cluster
configuration may exist on all nodes prior to the rabbitmq-server
package, we get into a state where the rabbitmq-server script is
partially available. But since we are sourcing the upstream script, when
it fails it returns 1 which is being interpreted as the wrong error.
This change returns not running if the script is not available since
clearly we can't be running if the package is not installed yet.

Change-Id: Idbc2a9ded39a47e06183793ac4a63115f93c9ba6
Closes-Bug: #1599479