Correctly return exit code from stop
Panicking and returning non-success on stop often leads to resource
becoming unmanaged on that node.

Before we called get_status to verify that RabbitMQ is dead. But
sometimes it returns error even though RabbitMQ is not running. There
is no reason to call it - we will just verify that there is no beam
process running.

Closes-Bug: #1626933
Change-Id: Iaeebc4fe3d29d06f8f94dfbb676cd30cf090f415