Update upload_cirros to support multiple images
This change creates a puppet parser function to convert the
test_vm_image from hiera into a glance_image compatible array of hashes
that can be used with create_resources. Previously vmware deployments
included an extra TestVM to be uploaded but during the rewrite of the
upload_cirros task this ability was dropped.

Change-Id: I3f0f6da32134a4a15f2c4764c5334bb8e85309c8
Closes-Bug: #1603393