Update pins for UCA Xenial
As we move to supporting xenial, we need to update the appropriate pins
for deploying with UCA. We no longer need to pin openvswitch and ceph as
they currently do not have packages to pin to. We are continuing to pin
our haproxy version and the rabbitmq versions as the MOS haproxy has the
conf.d patch we assume in fuel-library and the rabbitmq-server version
is newer than the upstream xenial version.

Change-Id: Ib55619d18d9a4268c4c5fa9252b721df6a3f2571
Closes-Bug: #1597721