[sync/mitaka] Set upper limit on keystone workers
This change sets an upper limit on the keystone workers to 16. It will
use the processor count to determine the correct number of workers for
smaller environments so it uses the smaller of the values between
processorcount or 16. This change sets the number of admin workers and
public workers to the lesser of these two values for a maximum of 32
workers total.

Change-Id: Ibbc2704441806776ded62e83b26b043188ad1a42
Closes-Bug: #1602434