Add refactored merge_yaml_settings module
Replace the merge_yaml_settings resource with the
merge_yaml module. This module introduced more advanced
version of the merge_yaml_settings resource with
additional options to control the merging behavior.

There are also resources with alternative implementation
of a configuration merging types and providers. They
are working in the same way as the "concat" module does
by assembling the resulting data structure from a set of
fragments. This should be much more puppet-master frinedly.
These types are not used anywhere yet but may be used later.

Related-Bug: 1614279
Cherry-picked-from: 9eccbbc
Change-Id: Ieaa8fc840f6a91a8d10e0670bd60d07692e2cb7d