[WIP] Add upgrade stuff for MU
If corosync package was upgraded it restarts corosync service
and it kills pacemaker service. So we need to start it again.
Also deny apt to stop pacemaker service during upgrade
because it leads to unload of all pacemaker resources
(all services under it will be stopped). That's not appropriate
behaviour during the update process. Puppet will manage services
restart only one by one.

Closes-bug: #1614893

Change-Id: I0d7231aa5900318f75f71c698f3e1c07f8e5cfbe