Add retry to master gtid query
If all the mysql nodes are booted at the exact same time, we can end up
with a situation where the master determination can occur almost at the
same time. This change updates the gtid fetching that is done during
master determination to include a retry with a random 1-10 second sleep
in an attempt to allow for the other nodes to update pacemaker with
their gtid information.

Change-Id: Ib12fb927391857ca9e3fb0a3ee45a7eec9e7913e
Closes-Bug: #1610180