Puppet4 support: fetch_value function
The functions for extracting values from a nested structure
have a long and funny history. Currently this function is
named "dig" and is used in the Fule linrary by this name.
But in Puppet 4.5 another incompatible "dig" function was
introduced and the original "dig" in the stdlib was renamed
to dig44.

The dig function from the stdlib doesn't support :undef values
and has problems with other values too. This function is
based on the "dig44" but has some improvements needed for
Puppet4.5+.

Related-Bug: 1586480
Fuel-CI: disable
Change-Id: Ieb4ae544d7ceab06c115a18df676cf8fa61e39a0