Use "safe" mirror list on upgrade
All "safe" repos for using in upgrade will be configured in UI and
available in Hiera. Then this list will be used to link this mirrors
to separate "sources.d" directory. This directory will be used to point
APT another source location. That allows us to get list of packages, which
will be upgraded with specified in "sources.d" mirrors. Then, available
packages will be upgraded.

To keep this process configurable we can specify hash of packages, which
should be upgraded throw hiera.

DocImpact

Closes-bug: #1616553
Closes-bug: #1621056

Change-Id: Idf4ff71c04f8a75c572bb24e560b3827f3d55bac