Use "safe" mirror list on upgrade
All "safe" repos for using in upgrade will be
configured in UI and available in Hiera.
This list will be used to link this mirrors
to separate "sources.d" directory. This directory
will be used to point APT another source location.
This allows us to run "apt-get dist-upgrade" with
known and verified list of package repositories.

To keep this process more safe "--no-remove"
option is added to "dist-upgrade" command.

Closes-bug: #1616553

Change-Id: Idf4ff71c04f8a75c572bb24e560b3827f3d55bac