compute: don't use 'host-model' CPU mode
The `host-model' CPU mode is known to result in non-working combinations
of CPU features and other parameters (such as CPUID level) which confuse
the guest OS. In particular on Core i7 hosts the avx2 instructions are
advertised to the guest kernel. However depending on the hardware, host
kernel version, *and* qemu these instructions might be unavailable in
the guest mode. Thus a guest kernel gets an invalid opcode exception
(and paincs) when trying to execute an avx2 instruction.

To avoid the problem never enable the `host-model' cpu_mode in nova.conf,
use the libvirt default instead.

As a side effect this patch should make it possible to use the (nested)
hardware assisted virtualization on virtual compute nodes (which might
make MOS/Fuel CI quite a bit faster).

Closes-Bug: #1618473
Change-Id: Id4d36d88ff452b91c2990af5de36b628a13d04ca