Move 'ceph_ready_check' task from shell to puppet
This change moves the 'ceph_ready_check' task from the shell to the
puppet to eliminate the extra point of the failure: the puppet shell
wrapper script for the task temporary created and uploaded to the
node by the astute. TODO: rewrite the ruby script to the custom
puppet function instead of use the exec.

Closes-Bug: 1619727
Change-Id: I725e6f2cbba3ab6252a4212b89f939fe05de5e14
Signed-off-by: Maksim Malchuk <mmalchuk@mirantis.com>