Upgrade a single package or a bunch of packages
It takes package names and versions from hiera hash to upgrade
and override "ensure" parameter in package resource.
If any package from the current puppet catalog is upgraded,
all the services in the same catalog should be restarted.

If there are no upgraded packages, then no services to be restarted.

Change-Id: Ic79cffb2b92cc0380261f5980263d84b7434a7f4
Closes-Bug: #1623038
Related-Bug: #1616553