Upgrade a single package or a bunch of packages
It takes package names and versions from hiera hash to upgrade
and override "ensure" parameter in package resource.
If any package from the current puppet catalog is upgraded,
all the services in the same catalog should be restarted.

If there are no upgraded packages, then no services to be restarted.

Change-Id: Ic3455c54aeea8890f8f3454c269d5f8a8ea0de4e
Closes-Bug: #1623038
Related-Bug: #1616553