Override package versions from hiera hash and notify all services in current catalog.
It takes package names and versions from hiera hash to upgrade and override "ensure" parameter in package resource
If any package from the current puppet catalog is upgraded, all the services from the same catalog should be restarted.
If there are no upgraded packages, than no services to be restarted.

Closes-bug: #1616553

Change-Id: Id99bfae07ea0a9e63ad319076743a6101f2675f8