Puppet4 support: refactor vips
Cherry-pick of change https://review.openstack.org/#/c/326541/

The generate_vips function is refactored to work in pair
with the create_resources functions. It's much better for testing
and debugging.
Input data of the vurual ip definition is split into many parameters
and the generate vips function is tought to support the changed
interfacs as well as to work with Puppet4's undef values.

Change-Id: I5fca1efaddeda3a8324ee5ff3b97ae9291b4e0dc
Related-Bug: 1586480