Split configuration astute.yaml into common and node parts
This file is huge if there are many nodes in cluster. Now
we generate it and store in memory for each node and it
takes a lot of time and memory.

Let's split it into common and node-specific parts so that
we can generate common part once and copy it from master
node to each node affected by deployment.

Change-Id: I929f9d0db80b7be01675697c972ac2be2273e182
Closes-Bug: #1596987