Ignore heartbeats lock fails
Stomp heartbeat handling is quite poorly designed. It happens in a
separate thread which sleeps, then tries to read a heartbeat if reading
mutex is acquired by message receiving thread it fails and increases
lock failure count. Upon reaching the limit (in our packets it is 2 by
default) it forcibly closes the connetion causing reconnect. Setting the
value to 0 turns the feature off.

Change-Id: Ieec889828d1dd2654ee760e7d5676efd14c7c348
Closes-Bug: #1613246
Closes-Bug: #1298262