Split configuration astute.yaml into common and node parts
This file is huge if there are many nodes in cluster. Now
we generate it and store in memory for each node and it
takes a lot of time and memory.

Let's split it into common and node-specific parts so that
we can generate common part once and copy it from master
node to each node affected by deployment.

Change-Id: I114a8598460bef9063e43505f46fe38cd564d8c7
Closes-Bug: #1660308