Add external hiera override hierarchy levels
Additional hierarchy levels can be added to
"/etc/hiera/external" with a r10k friendly structure.

These levels can be used to apply overrides from the
external repository either per-environment or for
the default production environemnt.

Overrides can be applies either for every node,
or for the roles group from the "roles" facts, or
globally and per-cluster.

Change-Id: Ia9e0bf6213ce616e96f7c3598a362e1978f48e2a
Related-Bug: 1618139
Rally-Id: US763570