Add the possibility to tune xinetd/tftpd server on Fuel master
This is a workaround for hanging tftpd processes.
The root cause can't be solved in Fuel, but this patch should ease
the load on the Fuel master.

The tftpd server is now started with an timeout option of 30 seconds
instead of the default 900 seconds.

The performance affecting parameters of xinetd & tftpd can be configured from
Fuel astute.yaml. The parameters are:
server_args : all options can be set for tftpd, '$args' as last asrgument
per_source  : This specifies the maximum instances of this service
              per source IP address (xinetd)
cps         : Limits the rate of incoming connections (xinetd)
This opens the possibility to have different parameters for different
size of environments.

Default configuration is equal to current, plus the timeout.

Change-Id: I00277b4e35e038edb36fe4f9c622e09723b675a5
Closes-bug: #1625716