Refactoring of the override_resources type for the IaC feature
This commit changes override_resources type allowing to
- specify any puppet resource
- pass any number of resources

It also introduces the logic to decide whether override_resources
should create a new resource, if it hasn't been found in the catalog.

The resource overrides as well as additional settings are taken from
Hiera.

This logic essentially allows to create any Puppet resource via
passing the right hash to the Hiera. It replaces/extends the old
'Advanced configuration' feature.

Change-Id: I81099ac41c3e2abdaf6fde3a19161709b7ba4b8c
Blueprint: puppet-resource-override