Split public_vip_ping task out
public_vip_ping task has been splitted out to avoid unwanted failover when
ping_checker resource might be started up on controllers before a primary one.
In addition pingd constraint has been changed to be based on score.
If all the nodes failed to ping the default gw, the resource
ain't stopped and remains on the last node where it was.

Change-Id: Ic02fc33f323f75c0b47e78e0935842dd9ae5a27d
Closes-Bug: #1643510