Stop configure TLS in Sahara internally
In our default architecture we use Haproxy to terminate TLS
for OpenStack services. If some plugin, driver or other service
needs to configure TLS for other service by other means, it should
be processed as additional feature and never as explicit enable of
TLS as it can break our domain model.

Change-Id: I0a35bba2a21665284d09f391e1e48058b4a5c64a
Closes-Bug: #1650284