Restart neutron services on neutron config change
There are many neutron tasks, e.g.: common-config and network-plugins-l2.
Normally the common-config configures neutron.conf and then plugns-l2 starts
the ovs-agent.

But there is the wrong behaviour when new controllers are added to an
environment:
  1. configure and deploy an env.
  2. common-config configures neutron.conf
  3. plugins-l2 starts ovs-agent.
  4. add new controllers to the env and redeploy it.
  5. common-config configures neutron.conf.
  6. plugins-l2 knows nothing about changes in the config.

So ovs-agent isn't restarted and works with old configuration data.

This commit makes neutron tasks check md5 hash of neutron.conf and
restart them when config is changed.

Change-Id: Ia26f4a5105407ce322a5185455093d69a938edfc
Closes-bug: #1613724