Restart ovs-agent when redeploy an environment
There are 2 tasks: common-config and network-plugins-l2. Normally the
common-config configures neutron.conf and then plugns-l2 starts
the ovs-agent.

But there is the wrong behaviour when new controllers are added to an
environment:
  1. configure and deploy an env.
  2. common-config configures neutron.conf
  3. plugins-l2 starts ovs-agent.
  4. add new controllers to the env and redeploy it.
  5. common-config configures neutron.conf.
  6. plugins-l2 knows nothing about changes in the config.

So ovs-agent doesn't restarted and works with old configuration data.

This commit changes the common-config task. If it changes neutron.conf
and ovs-agent init-script happens (i.e. in case of redeploy) the
ovs-agent would restarted.

Change-Id: Ia26f4a5105407ce322a5185455093d69a938edfc
Closes-bug: #1613724