Fix apt discrepancy around mysql client package
When we run galera/mysql packages installation
we can get into dependencies issues with  mysql
client package as it is not installed at the time
when perl dbd module requires it. At the same time
libperl-dbd-mysql is depending on it, so it fails.
Instead we need to ensure that mysql client is installed
along with its libraries prior to installing additional
packages for galera.

Change-Id: Idc91a399a46cd63c2201660bb248c0a0ad3fb9b8
Closes-bug: #1651182