Nailgun status check should include http responses
Fuel-utils could be used for automation of some tasks
and it is crucial service status checks to be adequate.
It is not enough to check that nailgun is started
we also need to check that it accepts requests and
responds.

Change-Id: I587925a468283cf48e22dc8afef36d7674af9903
Partial-Bug: #1656872
(cherry picked from commit 7c1608c2f4340b00779f2a0ea887d1f5294efead)