Remove log-error mysqld_safe option
According to [1], if file name has no extension,
the server adds an extension of ".err".

It causes two different issues.

First issue:
with log-error=/dev/stdout mysqld creates /dev/sdtout.err file
and redirects stderr to this file. This may cause OOM in case of intensive
logging, because the /dev is a devtmpfs which is in-memory pseudo file system.

Second issue:
the newer versions of mysql (5.6.35 at least) can't create
/dev/stdout.err file and fais with error.

[1] https://dev.mysql.com/doc/refman/5.6/en/server-options.html#option_mysqld_log-error

Change-Id: I430d8dd085433e8cb84fdf0595b8357758d9ea75